1.2017.01.05:解决第一次初始化失败的问题

+++ ChromeAutoUpdate/FrmMain.cs
@@ -524,7 +524,12 @@ namespace ChromeAutoUpdate
 
                 //移动chrome.exe
                 log(@"update\Chrome-bin\chrome.exe" + "到" + app_filename + @".new");
-                File.Move(@"update\Chrome-bin\chrome.exe", app_filename+@".new");
+
+                //如果存在，就保存新版本，不然就直接移动
+                if(File.Exists(app_filename))
+                    File.Move(@"update\Chrome-bin\chrome.exe", app_filename+@".new");
+                else
+                    File.Move(@"update\Chrome-bin\chrome.exe", app_filename);
 
                 try
                 {
